Win32-OpenSSH v9.5.0 modified files

+++ regress/test-exec.sh
@@ -40,34 +40,40 @@ fi
 unset SSH_AUTH_SOCK
 
 # Portable-specific settings.
-
-if [ -x /usr/ucb/whoami ]; then
-	USER=`/usr/ucb/whoami`
-elif whoami >/dev/null 2>&1; then
-	USER=`whoami`
-elif logname >/dev/null 2>&1; then
-	USER=`logname`
+if [ "x$TEST_WINDOWS_SSH" != "x" ]; then
+	os="windows"
+	USER=$TEST_SSH_USER
+	USER_DOMAIN=$TEST_SSH_USER_DOMAIN
+	LOGNAME=$USER
 else
-	USER=`id -un`
-fi
-if test -z "$LOGNAME"; then
-	LOGNAME="${USER}"
-	export LOGNAME
-fi
+	if [ -x /usr/ucb/whoami ]; then
+		USER=`/usr/ucb/whoami`
+	elif whoami >/dev/null 2>&1; then
+		USER=`whoami`
+	elif logname >/dev/null 2>&1; then
+		USER=`logname`
+	else
+		USER=`id -un`
+	fi
+	if test -z "$LOGNAME"; then
+		LOGNAME="${USER}"
+		export LOGNAME
+	fi
 
-# Unbreak GNU head(1)
-_POSIX2_VERSION=199209
-export _POSIX2_VERSION
+	# Unbreak GNU head(1)
+	_POSIX2_VERSION=199209
+	export _POSIX2_VERSION
 
-case `uname -s 2>/dev/null` in
-OSF1*)
-	BIN_SH=xpg4
-	export BIN_SH
-	;;
-CYGWIN*)
-	os=cygwin
-	;;
-esac
+	case `uname -s 2>/dev/null` in
+	OSF1*)
+		BIN_SH=xpg4
+		export BIN_SH
+		;;
+	CYGWIN*)
+		os=cygwin
+		;;
+	esac
+fi
 
 # If configure tells us to use a different egrep, create a wrapper function
 # to call it.  This means we don't need to change all the tests that depend
@@ -294,6 +300,21 @@ fi
 # to preserve our debug logging.  In the rare instance where -q is desirable
 # -qq is equivalent and is not removed.
 SSHLOGWRAP=$OBJ/ssh-log-wrapper.sh
+# BALU todo - check if we need to pass -T flag
+if [ "$os" == "windows" ]; then
+# timestamp line messes up stderr-data.sh stderr-after-eof.sh 
+# seems to be used for debugging concurrency tests (a feature unsupported on Windows currently)
+cat >$SSHLOGWRAP <<EOD
+#!/bin/sh
+logfile="${TEST_SSH_LOGDIR}/\${timestamp}.ssh.\$\$.log"
+echo "Executing: ${SSH} \$@" log \${logfile} >>$TEST_REGRESS_LOGFILE
+echo "Executing: ${SSH} \$@" >>\${logfile}
+for i in "\$@";do shift;case "\$i" in -q):;; *) set -- "\$@" "\$i";;esac;done
+rm -f $TEST_SSH_LOGFILE
+ln -f -s \${logfile} $TEST_SSH_LOGFILE
+exec ${SSH} -E\${logfile} "\$@"
+EOD
+else
 cat >$SSHLOGWRAP <<EOD
 #!/bin/sh
 timestamp="\`$OBJ/timestamp\`"
@@ -305,6 +326,7 @@ rm -f $TEST_SSH_LOGFILE
 ln -f -s \${logfile} $TEST_SSH_LOGFILE
 exec ${SSH} -E\${logfile} "\$@"
 EOD
+fi
 
 chmod a+rx $OBJ/ssh-log-wrapper.sh
 REAL_SSH="$SSH"
@@ -343,6 +365,9 @@ cat ${SSHAGENT_BIN} >${DATA}
 chmod u+w ${DATA}
 COPY=$OBJ/copy
 rm -f ${COPY}
+if [ "$os" == "windows" ]; then
+	EXEEXT=".exe"
+fi
 
 increase_datafile_size()
 {
@@ -357,6 +382,11 @@ export SSH_PKCS11_HELPER SSH_SK_HELPER
 #echo $SSH $SSHD $SSHAGENT $SSHADD $SSHKEYGEN $SSHKEYSCAN $SFTP $SFTPSERVER $SCP
 
 # Portable specific functions
+windows_path()
+{
+	cygpath -m $1
+}
+
 which()
 {
 	saved_IFS="$IFS"
@@ -428,36 +458,48 @@ fi
 
 make_tmpdir ()
 {
-	SSH_REGRESS_TMP="$($OBJ/mkdtemp openssh-XXXXXXXX)" || \
-	    fatal "failed to create temporary directory"
+	if [ "$os" == "windows" ]; then
+		powershell.exe /c "New-Item -Path $OBJ\openssh-XXXXXXXX -ItemType Directory -Force" >/dev/null 2>&1
+		if [ $? -ne 0 ]; then
+			fatal "failed to create temporary directory"
+		fi
+	else
+		SSH_REGRESS_TMP="$($OBJ/mkdtemp openssh-XXXXXXXX)" || \
+			fatal "failed to create temporary directory"
+	fi
 }
 # End of portable specific functions
 
 stop_sshd ()
 {
-	if [ -f $PIDFILE ]; then
-		pid=`$SUDO cat $PIDFILE`
-		if [ "X$pid" = "X" ]; then
-			echo no sshd running
-		else
-			if [ $pid -lt 2 ]; then
-				echo bad pid for sshd: $pid
+	# windows process can't be stopped using kill command so use stop-process
+	if [ "$os" == "windows" ]; then
+		powershell.exe /c "stop-process -Name sshd -Force" >/dev/null 2>&1
+	 else
+	 	if [ -f $PIDFILE ]; then
+			pid=`$SUDO cat $PIDFILE`
+			if [ "X$pid" = "X" ]; then
+				echo no sshd running
 			else
-				$SUDO kill $pid
-				trace "wait for sshd to exit"
-				i=0;
-				while [ -f $PIDFILE -a $i -lt 5 ]; do
-					i=`expr $i + 1`
-					sleep $i
-				done
-				if test -f $PIDFILE; then
-					if $SUDO kill -0 $pid; then
-						echo "sshd didn't exit " \
-						    "port $PORT pid $pid"
-					else
-						echo "sshd died without cleanup"
+				if [ $pid -lt 2 ]; then
+					echo bad pid for sshd: $pid
+				else
+					$SUDO kill $pid
+					trace "wait for sshd to exit"
+					i=0;
+					while [ -f $PIDFILE -a $i -lt 5 ]; do
+						i=`expr $i + 1`
+						sleep $i
+					done
+					if test -f $PIDFILE; then
+						if $SUDO kill -0 $pid; then
+							echo "sshd didn't exit " \
+								"port $PORT pid $pid"
+						else
+							echo "sshd died without cleanup"
+						fi
+						exit 1
 					fi
-					exit 1
 				fi
 			fi
 		fi
@@ -467,11 +509,19 @@ stop_sshd ()
 # helper
 cleanup ()
 {
-	if [ "x$SSH_PID" != "x" ]; then
-		if [ $SSH_PID -lt 2 ]; then
-			echo bad pid for ssh: $SSH_PID
-		else
-			kill $SSH_PID
+	# windows process can't be stopped using kill command so use stop-process
+	if [ "$os" == "windows" ]; then
+		powershell.exe /c "stop-process -Name ssh-agent -Force" >/dev/null 2>&1
+		if [ "x$SSH_PID" != "x" ]; then
+			powershell.exe /c "stop-process -Id $SSH_PID -Force" >/dev/null 2>&1
+		fi
+	else
+		if [ "x$SSH_PID" != "x" ]; then
+			if [ $SSH_PID -lt 2 ]; then
+				echo bad pid for ssh: $SSH_PID
+			else
+				kill $SSH_PID
+			fi
 		fi
 	fi
 	if [ "x$SSH_REGRESS_TMP" != "x" ]; then
@@ -596,6 +646,7 @@ cat << EOF > $OBJ/sshd_config
 	Subsystem	sftp	$SFTPSERVER
 EOF
 
+if [ "$os" != "windows" ]; then
 # This may be necessary if /usr/src and/or /usr/obj are group-writable,
 # but if you aren't careful with permissions then the unit tests could
 # be abused to locally escalate privileges.
@@ -626,6 +677,7 @@ bypass this check by setting TEST_SSH_UNSAFE_PERMISSIONS=1
 
 EOD
 	fi
+fi
 fi
 
 if [ ! -z "$TEST_SSH_MODULI_FILE" ]; then
@@ -699,6 +751,25 @@ maybe_filter_sk() {
 SSH_KEYTYPES=`$SSH -Q key-plain | maybe_filter_sk`
 SSH_HOSTKEY_TYPES=`$SSH -Q key-plain | maybe_filter_sk`
 
+if [ "$os" == "windows" ]; then
+	SSH_KEYTYPES=`echo $SSH_KEYTYPES | tr -d '\r','\n'`  # remove \r\n
+	SSH_HOSTKEY_TYPES=`echo $SSH_HOSTKEY_TYPES | tr -d '\r','\n'`  # remove \r\n
+	OBJ_WIN_FORMAT=`windows_path $OBJ`
+	first_key_type=${SSH_KEYTYPES%% *}
+	if [ "x$USER_DOMAIN" != "x" ]; then
+		# For domain user, create folders
+		if [ ! -d $OBJ/authorized_keys_$USER_DOMAIN ]; then
+			mkdir $OBJ/authorized_keys_$USER_DOMAIN
+		fi
+		if [ ! -d $OBJ/authorized_principals_$USER_DOMAIN ]; then
+			mkdir $OBJ/authorized_principals_$USER_DOMAIN
+		fi
+		if [ ! -d /var/run/principals_command_$USER_DOMAIN ]; then
+			mkdir /var/run/principals_command_$USER_DOMAIN
+		fi
+	fi
+fi
+
 for t in ${SSH_KEYTYPES}; do
 	# generate user key
 	if [ ! -f $OBJ/$t ] || [ ${SSHKEYGEN_BIN} -nt $OBJ/$t ]; then
@@ -724,12 +795,23 @@ for t in ${SSH_HOSTKEY_TYPES}; do
 
 	# use key as host key, too
 	(umask 077; $SUDO cp $OBJ/$t $OBJ/host.$t)
+	if [ "$os" == "windows" ]; then
+		# set the file permissions (ACLs) properly
+		pwsh.exe /ExecutionPolicy Bypass /c "get-acl $OBJ_WIN_FORMAT/$t | set-acl $OBJ_WIN_FORMAT/host.$t"
+		# powershell.exe /ExecutionPolicy Bypass /c "get-acl $OBJ_WIN_FORMAT/$t | set-acl $OBJ_WIN_FORMAT/host.$t"
+	fi
+
 	echo HostKey $OBJ/host.$t >> $OBJ/sshd_config
 
 	# don't use SUDO for proxy connect
 	echo HostKey $OBJ/$t >> $OBJ/sshd_proxy
 done
-chmod 644 $OBJ/authorized_keys_$USER
+
+if [ "$os" == "windows" ]; then
+	# set the file permissions (ACLs) properly
+	pwsh.exe /ExecutionPolicy Bypass /c "get-acl $OBJ_WIN_FORMAT/$first_key_type | set-acl $OBJ_WIN_FORMAT/authorized_keys_$USER"
+	# powershell.exe /ExecutionPolicy Bypass /c "get-acl $OBJ_WIN_FORMAT/$first_key_type | set-acl $OBJ_WIN_FORMAT/authorized_keys_$USER"
+fi
 
 # Activate Twisted Conch tests if the binary is present
 REGRESS_INTEROP_CONCH=no
@@ -789,7 +871,13 @@ fi
 # create a proxy version of the client config
 (
 	cat $OBJ/ssh_config
-	echo proxycommand ${SUDO} env SSH_SK_HELPER=\"$SSH_SK_HELPER\" ${OBJ}/sshd-log-wrapper.sh -i -f $OBJ/sshd_proxy
+	if [ "$os" == "windows" ]; then
+		# TODO - having SSH_SK_HELPER is causing issues. Need to find a way.
+		# This is fine for now as we don't have FIDO enabled.
+		echo proxycommand  `windows_path ${SSHD}` -i -f $OBJ_WIN_FORMAT/sshd_proxy
+	else
+		echo proxycommand ${SUDO} env SSH_SK_HELPER=\"$SSH_SK_HELPER\" ${OBJ}/sshd-log-wrapper.sh -i -f $OBJ/sshd_proxy
+	fi
 ) > $OBJ/ssh_proxy
 
 # check proxy config
@@ -800,8 +888,15 @@ start_sshd ()
 	# start sshd
 	logfile="${TEST_SSH_LOGDIR}/sshd.`$OBJ/timestamp`.$$.log"
 	$SUDO ${SSHD} -f $OBJ/sshd_config "$@" -t || fatal "sshd_config broken"
-	$SUDO env SSH_SK_HELPER="$SSH_SK_HELPER" \
-	    ${SSHD} -f $OBJ/sshd_config "$@" -E$TEST_SSHD_LOGFILE
+	if [ "$os" == "windows" ]; then
+		# In windows, we need to explicitly remove the sshd pid file.
+		rm -rf $PIDFILE
+		#TODO (Code BUG) : -E<sshd.log> is writing the data the cygwin terminal.
+		${SSHD} -f $OBJ/sshd_config "$@" &
+	else
+		$SUDO env SSH_SK_HELPER="$SSH_SK_HELPER" \
+	    	${SSHD} -f $OBJ/sshd_config "$@" -E$TEST_SSHD_LOGFILE
+	fi
 
 	trace "wait for sshd"
 	i=0;
